requirements.txt
django==3.2.8

gspread==4.0.1
oauth2client==4.1.3

python-dotenv >= 0.19.1
pyTelegramBotAPI >= 4.1.1

pymongo >= 3.12.1  # MongoDB not implemented at this moment

gunicorn >= 20.1.0
django-on-heroku >= 1.1.2  # Required for PostgreSQL implementation
psycopg2-binary >= 2.9.2  # PostgreSQL not implemented at this moment
